requirements.txt
torch
pyyaml
psutil
tensorly
pytest
transformers
datasets 